feat(Dropdown): add helper file
