+++ src/components/molecules/Dropdown/constants.ts
@@ -1,7 +1,6 @@
-export const DEFAULT_LOADING_TEXT = "Loading";
-export const DEFAULT_EMPTY_TEXT = "No data";
 export const DEFAULT_SEARCH_PLACEHOLDER = "Search";
 export const DEFAULT_SELECT_ALL_LABEL = "Select All";
 export const DEFAULT_CLEAR_LABEL = "Clear";
 export const MEASURE_SAFETY_OFFSET = 24;
 export const FALLBACK_VISIBLE_ITEMS = 2;
+export const MENU_GAP_FROM_TARGET = 4;
